requirements.txt
Django==1.8
djangorestframework==3.4.6
requests==2.9.1
python-dotenv